+++ tools/decode-status.py
@@ -139,8 +139,11 @@ a_setoption = [[
     "Reduced CT range for Alexa",
     "Use FriendlyNames instead of ShortAddresses when possible",
     "(AWS IoT) publish MQTT state to a device shadow",
+    "Enable Device Groups",
+    "PWM Dimmer Turn brightness LED's off 5 seconds after last change",
+    "PWM Dimmer Turn red LED on when powered off",
+    "PWM Dimmer Buttons control remote devices",
     "",
-    "","","","",
     "","","","",
     "","","","",
     "","","","",
@@ -193,7 +196,7 @@ a_features = [[
     "USE_ARDUINO_SLAVE","USE_HIH6","USE_HPMA","USE_TSL2591",
     "USE_DHT12","USE_DS1624","USE_GPS","USE_HOTPLUG",
     "USE_NRF24","USE_MIBLE","USE_HM10","USE_LE01MR",
-    "USE_AHT1x","USE_WEMOS_MOTOR_V1","",""
+    "USE_AHT1x","USE_WEMOS_MOTOR_V1","USE_DEVICE_GROUPS","USE_PWM_DIMMER"
     ],[
     "","","","",
     "","","","",
@@ -236,7 +239,7 @@ else:
         obj = json.load(fp)
 
 def StartDecode():
-    print ("\n*** decode-status.py v20200222 by Theo Arends and Jacek Ziolkowski ***")
+    print ("\n*** decode-status.py v20200227 by Theo Arends and Jacek Ziolkowski ***")
 
 #    print("Decoding\n{}".format(obj))